+++ src/design-system/components/core/index.ts
@@ -1,4 +1,5 @@
 export { default as AppBadge } from './AppBadge.vue';
+export { default as AppBrandLogo } from './AppBrandLogo.vue';
 export { default as AppButton } from './AppButton.vue';
 export { default as AppCard } from './AppCard.vue';
 export { default as AppInput } from './AppInput.vue';